+++ src/imio/actionspanel/__init__.py
@@ -3,7 +3,11 @@
 
 from zope.i18nmessageid import MessageFactory
 
+import logging
+
+
 ActionsPanelMessageFactory = MessageFactory('imio.actionspanel')
+logger = logging.getLogger("imio.actionspanel")
 
 
 def initialize(context):